requirements.txt
asgiref==3.5.2
Django==4.0.5
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
pytz==2022.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.1
